requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via starlette
click==8.1.7
    # via uvicorn
fastapi==0.109.2
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.6
    # via anyio
psycopg[binary]==3.1.18
    # via -r requirements.in
psycopg-binary==3.1.18
    # via psycopg
pydantic==2.6.1
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.16.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.25
    # via -r requirements.in
starlette==0.36.3
    # via fastapi
typing-extensions==4.9.0
    # via
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.27.0.post1
    # via -r requirements.in
